Harden compressed full-sync edge cases
Honor checksum policy on both ends of compressed diskless transfers, avoid retaining compressed sync RDBs as AOF bases, and preserve 64-bit dual-channel output accounting.

Add focused coverage for dual-channel capability fallback, receiver checksum bypass, and compressed AOF cleanup.

+++ src/childinfo.c
@@ -36,7 +36,7 @@ typedef struct {
     size_t cow;
     monotime cow_updated;
     double progress;
-    size_t repl_output_bytes;
+    uint64_t repl_output_bytes;
     childInfoType information_type; /* Type of information */
 } child_info_data;
 
@@ -65,7 +65,7 @@ void closeChildInfoPipe(void) {
 }
 
 /* Send save data to parent. */
-void sendChildInfoGeneric(childInfoType info_type, size_t keys, size_t repl_output_bytes, double progress, char *pname) {
+void sendChildInfoGeneric(childInfoType info_type, size_t keys, uint64_t repl_output_bytes, double progress, char *pname) {
     if (server.child_info_pipe[1] == -1) return;
 
     static monotime cow_updated = 0;
@@ -117,7 +117,7 @@ void sendChildInfoGeneric(childInfoType info_type, size_t keys, size_t repl_outp
 }
 
 /* Update Child info. */
-void updateChildInfo(childInfoType information_type, size_t cow, monotime cow_updated, size_t keys, size_t repl_output_bytes, double progress) {
+void updateChildInfo(childInfoType information_type, size_t cow, monotime cow_updated, size_t keys, uint64_t repl_output_bytes, double progress) {
     if (cow > server.stat_current_cow_peak) server.stat_current_cow_peak = cow;
 
     if (information_type == CHILD_INFO_TYPE_CURRENT_INFO) {
@@ -142,7 +142,7 @@ void updateChildInfo(childInfoType information_type, size_t cow, monotime cow_up
  * if complete data read into the buffer,
  * data is stored into *buffer, and returns 1.
  * otherwise, the partial data is left in the buffer, waiting for the next read, and returns 0. */
-int readChildInfo(childInfoType *information_type, size_t *cow, monotime *cow_updated, size_t *keys, size_t *repl_output_bytes, double *progress) {
+int readChildInfo(childInfoType *information_type, size_t *cow, monotime *cow_updated, size_t *keys, uint64_t *repl_output_bytes, double *progress) {
     /* We are using here a static buffer in combination with the server.child_info_nread to handle short reads */
     static child_info_data buffer;
     ssize_t wlen = sizeof(buffer);
@@ -177,7 +177,7 @@ void receiveChildInfo(void) {
     size_t cow;
     monotime cow_updated;
     size_t keys;
-    size_t repl_output_bytes;
+    uint64_t repl_output_bytes;
     double progress;
     childInfoType information_type;
